# backend/openapi-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AI Fight Club API",
    "description": "Complete API documentation for the AI Fight Club application - a platform for AI coding challenges and evaluation",
    "version": "1.0.0",
    "license": {
      "name": "ISC",
      "url": "https://opensource.org/licenses/ISC"
    },
    "contact": {
      "name": "Support",
      "url": "https://aifightclub.example.com",
      "email": "support@aifightclub.example.com"
    },
    "termsOfService": "https://aifightclub.example.com/terms",
    "x-api-versioning": {
      "strategy": "uri-path",
      "current": "v1",
      "deprecated": [],
      "sunset": []
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Current server (relative path)"
    }
  ],
...
  "paths": {
...
    "/users": {
      "get": {
        "summary": "Get all users",
        "description": "Retrieves a list of all users with pagination",
        "operationId": "getAllUsers",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/limitParam"
          },
          {
            "$ref": "#/components/parameters/offsetParam"
          },
          {
            "$ref": "#/components/parameters/sortParam"
          },
          {
            "$ref": "#/components/parameters/filterParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Users retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/User"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/PaginationParams"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "summary": "Get user by ID",
        "description": "Retrieves a user by their unique identifier",
        "operationId": "getUserById",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the user to retrieve",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login to the system",
        "description": "Authenticates a user and returns a JWT token",
        "operationId": "loginUser",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "User's email address"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "description": "User's password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "token": {
                          "type": "string",
                          "description": "JWT token for authentication"
                        },
                        "user": {
                          "$ref": "#/components/schemas/User"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid email or password"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/challenges": {
      "get": {
        "summary": "Get all challenges",
        "description": "Retrieves a list of all challenges with pagination",
        "operationId": "getAllChallenges",
        "tags": [
          "Challenges"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/limitParam"
          },
          {
            "$ref": "#/components/parameters/offsetParam"
          },
          {
            "$ref": "#/components/parameters/sortParam"
          },
          {
            "$ref": "#/components/parameters/filterParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Challenges retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Challenge"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/PaginationParams"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      },
      "post": {
        "summary": "Create a new challenge",
        "description": "Creates a new challenge in the system",
        "operationId": "createChallenge",
        "tags": [
          "Challenges"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "content",
                  "userEmail"
                ],
                "properties": {
                  "content": {
                    "type": "string",
                    "description": "Challenge content/question"
                  },
                  "userEmail": {
                    "type": "string",
                    "format": "email",
                    "description": "Email of the user the challenge is for"
                  },
                  "focusArea": {
                    "type": "string",
                    "description": "Focus area for the challenge"
                  },
                  "challengeType": {
                    "type": "string",
                    "description": "Type of challenge"
                  },
                  "formatType": {
                    "type": "string",
                    "description": "Format of the challenge"
                  },
                  "difficulty": {
                    "type": "string",
                    "enum": [
                      "beginner",
                      "intermediate",
                      "advanced",
                      "expert"
                    ],
                    "description": "Difficulty level"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Challenge created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Challenge"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
...
    "/evaluations": {
      "get": {
        "summary": "Get all evaluations",
        "description": "Retrieves a list of all evaluations with pagination",
        "operationId": "getAllEvaluations",
        "tags": [
          "Evaluations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/limitParam"
          },
          {
            "$ref": "#/components/parameters/offsetParam"
          },
          {
            "$ref": "#/components/parameters/sortParam"
          },
          {
            "$ref": "#/components/parameters/filterParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Evaluations retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Evaluation"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/PaginationParams"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      },
      "post": {
        "summary": "Create a new evaluation",
        "description": "Creates a new evaluation for a challenge",
        "operationId": "createEvaluation",
        "tags": [
          "Evaluations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "challengeId",
                  "userId",
                  "response"
                ],
                "properties": {
                  "challengeId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "ID of the challenge being evaluated"
                  },
                  "userId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "ID of the user who submitted the response"
                  },
                  "response": {
                    "type": "string",
                    "description": "User's response to evaluate"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Evaluation created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Evaluation"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token authentication. Use the token from the login response. Tokens expire after 24 hours."
      }
    },
    "parameters": {
      "limitParam": {
        "name": "limit",
        "in": "query",
        "description": "Maximum number of items to return per page (default: 20, max: 100)",
        "schema": {
          "type": "integer",
          "default": 20,
          "minimum": 1,
          "maximum": 100
        }
      },
      "offsetParam": {
        "name": "offset",
        "in": "query",
        "description": "Number of items to skip (for pagination)",
        "schema": {
          "type": "integer",
          "default": 0,
          "minimum": 0
        }
      },
      "sortParam": {
        "name": "sort",
        "in": "query",
        "description": "Field to sort by (prefix with - for descending order)",
        "schema": {
          "type": "string",
          "example": "-createdAt"
        }
      },
      "filterParam": {
        "name": "filter",
        "in": "query",
        "description": "Filter criteria in format field:operator:value (e.g., difficulty:eq:advanced)",
        "schema": {
          "type": "string"
        }
      }
    },
...
    "schemas": {
      "PaginationParams": {
        "type": "object",
        "description": "Parameters for pagination",
        "properties": {
          "limit": {
            "type": "integer",
            "default": 20,
            "minimum": 1,
            "maximum": 100,
            "description": "Maximum number of items to return"
          },
          "offset": {
            "type": "integer",
            "default": 0,
            "minimum": 0,
            "description": "Number of items to skip"
          },
          "sort": {
            "type": "string",
            "description": "Field to sort by (prefix with - for descending order)"
          }
        }
      },
...
    }
  }
}